chore: group category types
