--- packages/auth/src/types/enums/AuthResetPasswordStep.ts
@@ -1,7 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-export enum AuthResetPasswordStep {
-	CONFIRM_RESET_PASSWORD_WITH_CODE = 'CONFIRM_RESET_PASSWORD_WITH_CODE',
-	DONE = 'DONE'
-}
