--- packages/auth/src/types/enums/AuthSignUpStep.ts
@@ -1,10 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-/**
- * Denotes the next step in the Sign Up process.
- */
-export enum AuthSignUpStep {
-	CONFIRM_SIGN_UP = 'CONFIRM_SIGN_UP',
-	DONE = 'DONE',
-}
