+++ packages/auth/src/types/index.ts
@@ -4,38 +4,35 @@
 // TODO: Remove "./Auth" export
 export * from './Auth';
 
-// Enums
-export { AuthSignUpStep } from './enums/AuthSignUpStep';
-export { DeliveryMedium } from './enums/DeliveryMedium';
-export { AuthResetPasswordStep } from './enums/AuthResetPasswordStep';
-export { AuthSignInStep } from './enums/AuthSignInStep';
+export { AuthSignUpStep, AuthResetPasswordStep, AuthSignInStep } from './enums';
 
-// Models
-export { AdditionalInfo } from './models/AdditionalInfo';
-export { AnyAttribute } from './models/AnyAttribute';
-export { AuthCodeDeliveryDetails } from './models/AuthCodeDeliveryDetails';
-export { AuthNextSignUpStep } from './models/AuthNextSignUpStep';
-export { AuthStandardAttributeKey } from './models/AuthStandardAttributeKey';
-export { AuthUserAttributeKey } from './models/AuthUserAttributeKey';
-export { AuthUserAttribute } from './models/AuthUserAttribute';
-export { GetAttributeKey } from './models/GetAttributeKey';
-export { AuthNextResetPasswordStep } from './models/AuthNextResetPasswordStep';
-export { AuthNextSignInStep } from './models/AuthNextSignInStep';
+export {
+	AdditionalInfo,
+	DeliveryMedium,
+	AnyAttribute,
+	AuthCodeDeliveryDetails,
+	AuthNextSignUpStep,
+	AuthStandardAttributeKey,
+	AuthUserAttributeKey,
+	AuthUserAttribute,
+	GetAttributeKey,
+	AuthNextResetPasswordStep,
+	AuthNextSignInStep,
+} from './models';
 
+export { AuthServiceOptions, AuthSignUpOptions } from './options';
 
-// Options
-export { AuthServiceOptions } from './options/AuthServiceOptions';
-export { AuthSignUpOptions } from './options/AuthSignUpOptions';
-
-// Requests
-export { ConfirmResetPasswordRequest } from './requests/ConfirmResetPasswordRequest';
-export { ResetPasswordRequest } from './requests/ResetPasswordRequest';
-export { ResendSignUpCodeRequest } from './requests/ResendSignUpCodeRequest';
-
-export { SignUpRequest } from './requests/SignUpRequest';
-export { SignInRequest } from './requests/SignInRequest';
+export {
+	ConfirmResetPasswordRequest,
+	ResetPasswordRequest,
+	ResendSignUpCodeRequest,
+	SignUpRequest,
+	SignInRequest,
+} from './requests';
 
 // Results
-export { AuthSignUpResult } from './results/AuthSignUpResult';
-export { AuthSignInResult } from './results/AuthSignInResult';
-export { ResetPasswordResult } from './results/ResetPasswordResult';
+export {
+	AuthSignUpResult,
+	AuthSignInResult,
+	ResetPasswordResult,
+} from './results';
