--- packages/auth/src/types/models/AdditionalInfo.ts
@@ -1,7 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-/**
- * Additional data that may be returned from Auth APIs.
- */
-export type AdditionalInfo = { [key: string]: string };
