--- packages/auth/src/types/models/AnyAttribute.ts
@@ -1,10 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-/**
- * A custom user attribute type.
- *
- * @remarks
- * This type can be used to represent use attributes not captured by standard OIDC claims.
- */
-export type AnyAttribute = (string & {}) | Record<string, string>;
