+++ packages/auth/src/types/options.ts
@@ -1,11 +1,12 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import {
-	AuthServiceOptions,
-	AuthUserAttribute,
-	AuthUserAttributeKey,
-} from '..';
+import { AuthUserAttribute, AuthUserAttributeKey } from "./models";
+
+/**
+ * Base type for service options.
+ */
+export type AuthServiceOptions = any;
 
 /**
  * The optional parameters for the Sign Up process.
