--- packages/auth/src/types/options/AuthServiceOptions.ts
@@ -1,7 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-/**
- * Base type for service options.
- */
-export type AuthServiceOptions = any;
